moved visit related class to target package
