+++ modules/web/src/com/haulmont/sample/petclinic/web-app.properties
@@ -45,4 +45,4 @@ cuba.gui.genericFilterApplyImmediately=false
 cuba.rest.securityScope=GENERIC_UI
 
 
-cuba.iconsConfig = +com.haulmont.sample.petclinic.web.visit.visit.VisitTypeIcon
+cuba.iconsConfig = +com.haulmont.sample.petclinic.web.screens.visit.VisitTypeIcon
