+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/visit/CalendarScreenAdjustment.java
@@ -1,4 +1,4 @@
-package com.haulmont.sample.petclinic.web.visit.visit.calendar.navigation;
+package com.haulmont.sample.petclinic.web.screens.visit;
 
 import com.haulmont.cuba.gui.components.Calendar;
 import com.haulmont.cuba.gui.components.DatePicker;
