+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/visit/RelativeDates.java
@@ -1,4 +1,4 @@
-package com.haulmont.sample.petclinic.web.visit.visit.calendar;
+package com.haulmont.sample.petclinic.web.screens.visit;
 
 import java.time.*;
 import java.time.temporal.TemporalAdjusters;
