+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/visit/VisitTypeIcon.java
@@ -1,4 +1,4 @@
-package com.haulmont.sample.petclinic.web.visit.visit;
+package com.haulmont.sample.petclinic.web.screens.visit;
 
 import com.haulmont.cuba.gui.icons.Icons;
