+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/visit/messages.properties
@@ -1,0 +1,17 @@
+calender=Calender
+table=Table
+type=Visit Type
+filter=Filter
+calendarMode=Calendar Mode
+overview=Overview
+browseCaption = Visits
+editorCaption = Visit
+createForPet=Create Visit for Pet
+caption = Create visit for Pet
+petId=Pet ID
+create=Create
+petIdRequired=Pet ID required
+visitUpdated='%s' Visit updated for %s
+CalendarMode.DAY = 1
+CalendarMode.WEEK = 7
+CalendarMode.MONTH = 30
