+++ modules/web/src/com/haulmont/sample/petclinic/web/screens/visit/visit-edit.xml
@@ -2,7 +2,7 @@
 <window xmlns="http://schemas.haulmont.com/cuba/screen/window.xsd"
         caption="msg://editorCaption"
         focusComponent="fieldGroup"
-        messagesPack="com.haulmont.sample.petclinic.web.visit.visit">
+        messagesPack="com.haulmont.sample.petclinic.web.screens.visit">
     <data>
         <instance id="visitDc"
                   class="com.haulmont.sample.petclinic.entity.visit.Visit"
