--- modules/web/src/com/haulmont/sample/petclinic/web/visit/messages.properties
@@ -1,6 +1,0 @@
-calender=Calender
-table=Table
-type=Visit Type
-filter=Filter
-calendarMode=Calendar Mode
-overview=Overview
