--- modules/web/src/com/haulmont/sample/petclinic/web/visit/visit/calendar/messages.properties
@@ -1,3 +1,0 @@
-CalendarMode.DAY = 1
-CalendarMode.WEEK = 7
-CalendarMode.MONTH = 30
