--- modules/web/src/com/haulmont/sample/petclinic/web/visit/visit/messages.properties
@@ -1,8 +1,0 @@
-browseCaption = Visits
-editorCaption = Visit
-createForPet=Create Visit for Pet
-caption = Create visit for Pet
-petId=Pet ID
-create=Create
-petIdRequired=Pet ID required
-visitUpdated='%s' Visit updated for %s
